renamed dependencies
