+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,6 +1,6 @@
 ---
 blank_issues_enabled: false
 contact_links:
-  - name: Madara Community Support
-    url: https://github.com/keep-starknet-strange/madara/discussions
+  - name: Deoxys Community Support
+    url: https://github.com/kasarlabs/deoxys/discussions
     about: Please ask and answer questions here.
