+++ crates/client/rpc/src/events/mod.rs
@@ -12,7 +12,7 @@ use starknet_core::types::{BlockId, BlockTag, EmittedEvent};
 use starknet_ff::FieldElement;
 
 use crate::errors::StarknetRpcApiError;
-use crate::madara_backend_client::get_block_by_block_hash;
+use crate::deoxys_backend_client::get_block_by_block_hash;
 use crate::Starknet;
 
 impl<BE, C, H> Starknet<BE, C, H>
