+++ crates/client/rpc/src/methods/get_block.rs
@@ -13,7 +13,7 @@ use starknet_core::types::{
     PendingBlockWithTxs,
 };
 
-use crate::madara_backend_client::get_block_by_block_hash;
+use crate::deoxys_backend_client::get_block_by_block_hash;
 use crate::utils::block::{
     l1_da_mode, l1_data_gas_price, l1_gas_price, new_root, parent_hash, sequencer_address, starknet_version, timestamp,
 };
