+++ crates/client/rpc/src/methods/read/call.rs
@@ -11,7 +11,7 @@ use starknet_api::transaction::Calldata;
 use starknet_core::types::{BlockId, FunctionCall};
 
 use crate::errors::StarknetRpcApiError;
-use crate::madara_backend_client::get_block_by_block_hash;
+use crate::deoxys_backend_client::get_block_by_block_hash;
 use crate::{utils, Arc, Starknet};
 
 /// Call a Function in a Contract Without Creating a Transaction
