+++ crates/client/rpc/src/methods/read/estimate_fee.rs
@@ -14,7 +14,7 @@ use starknet_core::types::{
 };
 
 use crate::errors::StarknetRpcApiError;
-use crate::madara_backend_client::get_block_by_block_hash;
+use crate::deoxys_backend_client::get_block_by_block_hash;
 use crate::{utils, Starknet};
 
 /// Estimate the fee associated with transaction
@@ -77,7 +77,7 @@ where
 
     let estimates = fee_estimates
             .into_iter()
-			// FIXME: https://github.com/keep-starknet-strange/madara/issues/329
+			// FIXME: https://github.com/kasarlabs/deoxys/issues/329
             // TODO: reflect right estimation
             .map(|x| FeeEstimate { gas_consumed: x.gas_consumed , gas_price: x.gas_price, data_gas_consumed: x.data_gas_consumed, data_gas_price: x.data_gas_price, overall_fee: x.overall_fee, unit: x.unit})
             .collect();
