+++ crates/client/rpc/src/methods/read/get_block_transaction_count.rs
@@ -9,7 +9,7 @@ use sp_blockchain::HeaderBackend;
 use starknet_core::types::BlockId;
 
 use crate::errors::StarknetRpcApiError;
-use crate::madara_backend_client::get_block_by_block_hash;
+use crate::deoxys_backend_client::get_block_by_block_hash;
 use crate::Starknet;
 
 /// Get the Number of Transactions in a Given Block
