+++ crates/client/rpc/src/methods/read/get_state_update.rs
@@ -15,7 +15,7 @@ use starknet_api::block::BlockHash as APIBlockHash;
 use starknet_core::types::{BlockId, BlockTag, FieldElement, MaybePendingStateUpdate, StateDiff, StateUpdate};
 
 use crate::errors::StarknetRpcApiError;
-use crate::madara_backend_client::get_block_by_block_hash;
+use crate::deoxys_backend_client::get_block_by_block_hash;
 use crate::Starknet;
 
 fn get_state_update_finalized<BE, C, H>(
