+++ crates/client/rpc/src/methods/read/get_transaction_status.rs
@@ -13,7 +13,7 @@ use sp_blockchain::HeaderBackend;
 use starknet_core::types::{FieldElement, TransactionExecutionStatus, TransactionStatus};
 
 use crate::errors::StarknetRpcApiError;
-use crate::madara_backend_client::get_block_by_block_hash;
+use crate::deoxys_backend_client::get_block_by_block_hash;
 use crate::Starknet;
 
 /// Gets the Transaction Status, Including Mempool Status and Execution Details
