+++ crates/client/rpc/src/methods/trace/trace_block_transactions.rs
@@ -12,7 +12,7 @@ use starknet_core::types::{BlockId, TransactionTraceWithHash};
 
 use super::utils::{map_transaction_to_user_transaction, tx_execution_infos_to_tx_trace};
 use crate::errors::StarknetRpcApiError;
-use crate::madara_backend_client::get_block_by_block_hash;
+use crate::deoxys_backend_client::get_block_by_block_hash;
 use crate::methods::trace::utils::block_number_by_id;
 use crate::utils::execution::re_execute_transactions;
 use crate::Starknet;
